requirements.txt
# Core Flask and Extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Session==0.5.0
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.1
Werkzeug==3.0.1
cryptography==41.0.7

# OAuth
requests==2.31.0
oauthlib==3.2.2
requests-oauthlib==1.3.1

# Email (for magic links)
Flask-Mail==0.9.1

# Environment & Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
Flask-DebugToolbar==0.14.1
black==23.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0
